requirements.txt
fastapi>=0.110.0,<0.111.0
uvicorn>=0.25.0,<0.26.0
llama-index>=0.11.0,<0.12.0
llama-index-vector-stores-chroma>=0.2.0,<0.3.0
openai>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.6.0,<3.0.0
chromadb>=0.4.24,<0.5.0
typing-extensions>=4.11.0,<5.0.0 